% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationParms.R
\name{link2input}
\alias{link2input}
\title{Convert Browser Location Parameters to Shiny Input Values}
\usage{
link2input(..., inputtype = "default")
}
\arguments{
\item{...}{List of Shiny input IDs to match with window location parameters}

\item{inputtype}{Type of inputs being included}
}
\value{
Setting of the Shiny inputs to the values of the parameters in the browser link
}
\description{
Parse the browser link and retrieve parameters for inclusion as values in inputs
}
\note{
the entered inputs should be a list in order to be parsed
}
\examples{

link2input(id1 = "name", id2 = "type")
}
