% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeex.R
\name{seeexample}
\alias{seeexample}
\title{Load the example for the package}
\usage{
seeexample()
}
\value{
An example of inputs persistently stored when changed and the page refreshed
}
\description{
Example of a shiny application with secure in-browser storage of inputs when changed
}
\note{
Changes made to the input will be saved and returned when the page is refresh within the same browser over different sessions
}
\examples{
\donttest{
if (interactive()) {
  seeexample()
}
}

}
