% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyStorePlus-package.R
\docType{package}
\name{shinyStorePlus-package}
\alias{shinyStorePlus}
\alias{shinyStorePlus-package}
\title{shinyStorePlus: Secure in-Browser Storage for 'Shiny' Inputs and Variables}
\description{
Store persistent and synchronized data from 'Shiny' inputs within the browser in a secure format. Refresh 'Shiny' applications and preserve user-inputs over multiple sessions. A database-like storage format is implemented using 'Dexie.js' \url{https://dexie.org}, a minimal wrapper for 'IndexedDB'. Transfer browser link parameters to 'Shiny' input or output values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oobianom/shinyStorePlus}
  \item Report bugs at \url{https://github.com/oobianom/shinyStorePlus}
}

}
\author{
\strong{Maintainer}: Obinna Obianom \email{idonshayo@gmail.com}

}
\keyword{internal}
