% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeex.R
\name{seeexample}
\alias{seeexample}
\title{Load the example for the package}
\usage{
seeexample(name = c("storeInputs", "browserLinkToInput"))
}
\arguments{
\item{name}{the name of example to view. choices include storeInputs or browserLinkToInput}
}
\value{
An example of inputs persistently stored when changed and the page refreshed
}
\description{
Example of a shiny application with secure in-browser storage of inputs when changed
}
\note{
Changes made to the input or putputs will be saved and returned when the page is refresh within the same browser over different sessions. More examples are located at https://github.com/oobianom/aagarw30_shinyapps_to-shinyStorePlus
}
\examples{
\donttest{
if (interactive()) {
  seeexample()
  seeexample("browserLinkToInput")
}
}

}
