% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyStorePlus-package.R
\docType{package}
\name{shinyStorePlus-package}
\alias{shinyStorePlus}
\alias{shinyStorePlus-package}
\title{shinyStorePlus: Secure in-Browser and Database Storage for 'shiny' Inputs, Outputs, Views and User Likes}
\description{
Store persistent and synchronized data from 'shiny' inputs within the browser. Refresh 'shiny' applications and preserve user-inputs over multiple sessions. A database-like storage format is implemented using 'Dexie.js' \url{https://dexie.org}, a minimal wrapper for 'IndexedDB'. Transfer browser link parameters to 'shiny' input or output values. Store app visitor views, likes and followers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shinystoreplus.obi.obianom.com}
  \item Report bugs at \url{https://github.com/oobianom/shinyStorePlus/issues}
}

}
\author{
\strong{Maintainer}: Obinna Obianom \email{idonshayo@gmail.com}

}
\keyword{internal}
